# src/app/api/chatgpt/openapi.yaml
openapi: 3.1.0
info:
  title: Pandora's Box Memory API
  description: API for ChatGPT to store and retrieve memories from Pandora's Box
  version: 1.0.0
servers:
  - url: https://studio-sg--seismic-vista-480710-q5.asia-southeast1.hosted.app
    description: Production server
paths:
  /api/chatgpt/store-memory:
    post:
      operationId: storeMemory
      summary: Store a memory in Pandora's Box
      description: Stores a memory that can be retrieved later using semantic search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - memory
              properties:
                memory:
                  type: string
                  description: The memory content to store (e.g., "User prefers dark mode", "User's favorite programming language is TypeScript")
                user_email:
                  type: string
                  description: User email address (defaults to DEFAULT_CHATGPT_USER_EMAIL env var if not provided)
                  default: user@example.com
            examples:
              example1:
                value:
                  memory: "User prefers dark mode interface"
                  user_email: "user@example.com"
              example2:
                value:
                  memory: "User is working on a Next.js project with Firebase App Hosting"
                  user_email: "user@example.com"
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Memory stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  memory_id:
                    type: string
                  user_id:
                    type: string
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Server error

  /api/chatgpt/retrieve-memories:
    get:
      operationId: retrieveMemories
      summary: Retrieve relevant memories from Pandora's Box
      description: Retrieves memories using semantic search or returns recent memories
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Search query for semantic memory retrieval
        - name: user_email
          in: query
          required: false
          schema:
            type: string
          description: User email address (defaults to joven.ong23@gmail.com)
          default: joven.ong23@gmail.com
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: Maximum number of memories to return
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Memories retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  memories:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        content:
                          type: string
                        relevance_score:
                          type: number
                          description: Relevance score (0-1, higher is more relevant)
                        timestamp:
                          type: string
                          format: date-time
                  user_id:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Server error
    post:
      operationId: retrieveMemoriesPost
      summary: Retrieve relevant memories from Pandora's Box (POST)
      description: Alternative POST endpoint for retrieving memories
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: Search query for semantic memory retrieval
                user_email:
                  type: string
                  description: User email address (defaults to joven.ong23@gmail.com)
                  default: joven.ong23@gmail.com
                limit:
                  type: integer
                  minimum: 1
                  maximum: 50
                  default: 10
                  description: Maximum number of memories to return
            examples:
              example1:
                value:
                  query: "user preferences"
                  user_email: "user@example.com"
                  limit: 5
              example2:
                value:
                  query: "current project"
                  limit: 10
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Memories retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  memories:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        content:
                          type: string
                        relevance_score:
                          type: number
                        timestamp:
                          type: string
                          format: date-time
                  user_id:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Server error

components:
  schemas: {}
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key in format "Bearer YOUR_API_KEY"